test(util): 添加了applyAttrs的单元测试
